requirements.txt
discord.py[voice]
pillow
numpy
scipy
requests
speedtest-cli
psutil
asyncio
youtube_dl